Add e2e tests (#23)
* Add e2e tests

Goal: Validate that the FE & BE are up-and-running together

Process:
Navigate to a route that displays pre-seeded data from the API

This confirms:
- That the API is running (can be built successfully)
- That the FE is running (can be built successfully)
- That the FE is successfully hitting the API (the FE/BE have not departed)

* Include healthcheck in docker-compose

According to the documentation for docker-compose version 3, a
healthcheck should delay any dependent services until it passes.

* Add retry logic to e2e API test

* Add e2e tests badge

I'm not sure why Run Unit Tests is all capitalized and Run e2e tests is
not. They appear to be configured the same in the code.